+++ src/classes/job.ts
@@ -23,6 +23,23 @@ export interface JobJson {
   failedReason: string;
   stacktrace: string;
   returnvalue: string;
+  parentKey?: string;
+}
+
+export interface JobJsonRaw {
+  id: string;
+  name: string;
+  data: string;
+  opts: string;
+  progress: string;
+  attemptsMade: string;
+  finishedOn?: string;
+  processedOn?: string;
+  timestamp: string;
+  failedReason: string;
+  stacktrace: string[];
+  returnvalue: string;
+  parentKey?: string;
 }
 
 export class Job<T = any, R = any, N extends string = string> {
@@ -119,13 +136,13 @@ export class Job<T = any, R = any, N extends string = string> {
     return jobInstances;
   }
 
-  static fromJSON(queue: MinimalQueue, json: any, jobId?: string) {
+  static fromJSON(queue: MinimalQueue, json: JobJsonRaw, jobId?: string) {
     const data = JSON.parse(json.data || '{}');
     const opts = JSON.parse(json.opts || '{}');
 
     const job = new Job(queue, json.name, data, opts, json.id || jobId);
 
-    job.progress = JSON.parse(json.progress || 0);
+    job.progress = JSON.parse(json.progress || '0');
 
     // job.delay = parseInt(json.delay);
     job.timestamp = parseInt(json.timestamp);
@@ -139,7 +156,7 @@ export class Job<T = any, R = any, N extends string = string> {
     }
 
     job.failedReason = json.failedReason;
-    job.attemptsMade = parseInt(json.attemptsMade || 0);
+    job.attemptsMade = parseInt(json.attemptsMade || '0');
 
     job.stacktrace = getTraces(json.stacktrace);
 
@@ -162,7 +179,9 @@ export class Job<T = any, R = any, N extends string = string> {
     if (jobId) {
       const client = await queue.client;
       const jobData = await client.hgetall(queue.toKey(jobId));
-      return isEmpty(jobData) ? undefined : Job.fromJSON(queue, jobData, jobId);
+      return isEmpty(jobData)
+        ? undefined
+        : Job.fromJSON(queue, (<unknown>jobData) as JobJsonRaw, jobId);
     }
   }
 
@@ -247,7 +266,7 @@ export class Job<T = any, R = any, N extends string = string> {
     returnValue: R,
     token: string,
     fetchNext = true,
-  ): Promise<[JobJson, string] | []> {
+  ): Promise<[JobJsonRaw, string] | []> {
     await this.queue.waitUntilReady();
 
     this.returnvalue = returnValue || void 0;
@@ -570,7 +589,7 @@ export class Job<T = any, R = any, N extends string = string> {
   }
 }
 
-function getTraces(stacktrace: any[]) {
+function getTraces(stacktrace: string[]) {
   const traces = tryCatch(JSON.parse, JSON, [stacktrace]);
 
   if (traces === errorObject || !(traces instanceof Array)) {