+++ src/test/test_rate_limiter.ts
@@ -74,7 +74,56 @@ describe('Rate Limiter', function() {
       },
     });
 
-    const result = new Promise((resolve, reject) => {
+    const result = new Promise<void>((resolve, reject) => {
+      queueEvents.on(
+        'completed',
+        // after every job has been completed
+        after(numJobs, async () => {
+          await worker.close();
+
+          try {
+            const timeDiff = new Date().getTime() - startTime;
+            expect(timeDiff).to.be.gte((numJobs - 1) * 1000);
+            resolve();
+          } catch (err) {
+            reject(err);
+          }
+        }),
+      );
+
+      queueEvents.on('failed', async err => {
+        await worker.close();
+        reject(err);
+      });
+    });
+
+    for (let i = 0; i < numJobs; i++) {
+      await queue.add('rate test', {});
+    }
+
+    await result;
+    await worker.close();
+    await queueScheduler.close();
+  });
+
+  it('should obey the rate limit with workerDelay enabled', async function() {
+    this.timeout(20000);
+
+    const numJobs = 4;
+    const startTime = new Date().getTime();
+
+    const queueScheduler = new QueueScheduler(queueName);
+    await queueScheduler.waitUntilReady();
+
+    const worker = new Worker(queueName, async job => {}, {
+      limiter: {
+        max: 1,
+        duration: 1000,
+        workerDelay: true,
+      },
+    });
+
+    const result = new Promise<void>((resolve, reject) => {
       queueEvents.on(
         'completed',
         // after every job has been completed
@@ -132,7 +181,7 @@ describe('Rate Limiter', function() {
 
     const completed: { [index: string]: number[] } = {};
 
-    const running = new Promise((resolve, reject) => {
+    const running = new Promise<void>((resolve, reject) => {
       const afterJobs = after(numJobs, () => {
         try {
           const timeDiff = Date.now() - startTime;
@@ -236,7 +285,7 @@ describe('Rate Limiter', function() {
 
     await queue.resume();
 
-    const result = new Promise((resolve, reject) => {
+    const result = new Promise<void>((resolve, reject) => {
       queueEvents.on('failed', async err => {
         await worker.close();
         reject(err);